## Fixes
- Fixed missing cast when converting to radians.

+++ src/com/projectkorra/projectkorra/GeneralMethods.java
@@ -848,7 +848,7 @@ public class GeneralMethods {
 	 * @return the location of the distance to the left
 	 */
 	public static Location getLeftSide(final Location location, final double distance) {
-		final float angle = Math.toRadians(location.getYaw());
+		final float angle = (float) Math.toRadians(location.getYaw());
 		return location.clone().add(new Vector(Math.cos(angle), 0, Math.sin(angle)).normalize().multiply(distance));
 	}
 
@@ -898,7 +898,7 @@ public class GeneralMethods {
 	 * @return the location of the distance to the right
 	 */
 	public static Location getRightSide(final Location location, final double distance) {
-		final float angle = Math.toRadians(location.getYaw());
+		final float angle = (float) Math.toRadians(location.getYaw());
 		return location.clone().subtract(new Vector(Math.cos(angle), 0, Math.sin(angle)).normalize().multiply(distance));
 	}
